GLM-58 Progress and client request rework So this issue ended up being far more complex than initially estimated. The first problem encountered was the fact that Unity by default url-encodes the body of your web requests which means that the body-parser wont be able to read the bodies properly. This has been fixed by not using any of Unity's own WebRequest helper methods and instead build the requests programatically to circumvent the issue.
The next issue that cropped up is the fact that GET requests dont seem to work at the moment.
After some research it seems like the body-parser will completely ignore bodies for GET requests as these are "supposed" to be without a body.
The annoying part of this is the fact that the HTTP specification states that GET requests actually can have bodies, but this is not enforced so middleware and so on are free to ignore it.

 #time 3h 5m