GLM-91 Fix json library given invalid strings
Fixed the issue that arose if the json library was given invalid
strings.
This happened because curl calls your callback continously, which I had
missunderstood. Doing conversions in the callback is therefore not
advicable as you might not get all the information you need.
To fix this I simply moved all the parsing out of the callbacks and only
store the data in there. That also has the nice property of allowing all
the functions to use the same callbacks, which they do now.

 #time 1h 20m
 #partial-finish